+++ ITEIT87x/ITEIT87x.h
@@ -113,11 +113,11 @@ private:
     virtual bool			probePort();
     virtual bool			startPlugin();
 	
-	virtual long			readTemperature(unsigned long index);
-	virtual long			readVoltage(unsigned long index);
-	virtual long			readTachometer(unsigned long index);
+	virtual UInt16			readTemperature(UInt32 index);
+	virtual UInt16			readVoltage(UInt32 index);
+	virtual UInt16			readTachometer(UInt32 index);
 	
-    virtual int				getPortsCount();
+    virtual UInt8			getPortsCount();
 	virtual const char *	getModelName();
 	
 public:
