+++ NCT677x/NuvotonNCT677x.h
@@ -142,9 +142,9 @@ private:
     
     //void					updateTemperatures();
     
-	virtual long			readTemperature(unsigned long index);
-	virtual long			readVoltage(unsigned long index);
-	virtual long			readTachometer(unsigned long index);
+	virtual UInt16			readTemperature(UInt32 index);
+	virtual UInt16			readVoltage(UInt32 index);
+	virtual UInt16			readTachometer(UInt32 index);
 	
 	virtual const char *	getModelName();
